+++ sheerwater_benchmarking/reanalysis/era5.py
@@ -260,7 +260,7 @@ def era5_agg(start_time, end_time, variable, grid="global1_5", agg=14, mask="lsm
            timeseries='time',
            cache=False,
            cache_args=['variable', 'lead', 'grid', 'mask', 'region'])
-def era5(start_time, end_time, variable, lead, grid='africa0_25', mask='lsm', region='africa'):
+def era5(start_time, end_time, variable, lead, grid='global0_25', mask='lsm', region='africa'):
     """Standard format task data for ERA5 Reanalysis.
 
     Args:
