+++ sheerwater_benchmarking/utils/space_utils.py
@@ -3,6 +3,8 @@ import numpy as np
 import xarray as xr
 import geopandas as gpd
 
+from .general_utils import load_object
+
 
 def get_grid_ds(grid_id, base="base180"):
     """Get a dataset equal to ones for a given region."""
